requirements.txt
flask==3.0.0
google-generativeai==0.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
langchain
langchain_community
chromadb
langchain_chroma
sentence-transformers
langchain-ollama
pypdf